requirements.txt
# Core ML/DL frameworks
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.26.0

# Vision and language model specific
open-clip-torch>=2.20.0
vllm>=0.4.0
openai>=1.0.0

# Computer vision
Pillow>=10.0.0
pycocotools>=2.0.7
torchvision>=0.15.0

# Data processing and analysis
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
ipython>=8.0.0

# Utilities
tqdm>=4.65.0
fast-pytorch-kmeans>=0.2.0
pyarrow>=12.0.0
faiss-cpu>=1.7.4  # Use faiss-gpu if CUDA available

# Optional but recommended
jupyter>=1.0.0
notebook>=7.0.0